requirements.txt
openai
python-dotenv
tk
requests
functions_framework
google-cloud-storage
google-auth